feat: group signing/keystore modules under dedicated path
