+++ packages/client/src/index.ts
@@ -2,6 +2,6 @@ export * from './crypto';
 export * from './funded_account';
 export * from './interfaces';
 export * from './providers';
-export * from './signers';
+export * from './signing';
 export * from './transactions';
 export * from './view';
