+++ packages/client/src/signing/index.ts
@@ -1,0 +1,7 @@
+export {
+  getPlaintextFilesystemSigner,
+} from './keystores/plaintext_filesystem';
+export {
+  getSignerFromKeyPair,
+  getSignerFromPrivateKey,
+} from './signers';
